separated Categories, FeaturedProducts, RecentlyListedProducts from HomePage.jsx and Modified Featured and RecentlyListed to accomodate UI if Products are less than 3
